Mentioned issues are fixed
