+++ app/(authenticated)/leads/page.tsx
@@ -1,12 +1,8 @@
-import LeadDetails from "@/components/ui/LeadDetails/lead-details";
-import LeadsList from "@/components/ui/LeadsList/LeadsList";
+import LeadsList from "@/components/LeadsList/LeadsList";
 
 const LeadsPage = () => {
   return (
-    <div className=''>
       <LeadsList/>
-      {/* <LeadDetails/> */}
-    </div>
   );
 };
