+++ components/createlead/create-lead-form.tsx
@@ -1,12 +1,12 @@
 'use client';
 import Image from 'next/image';
 import mapImage from '@/assets/images/dummy-map.png';
-import { Input } from '../ui/input';
-import { TextArea } from '../ui/text-area';
-import { ImageUpload } from '../ui/image-upload';
-import { Button } from '../ui/button';
-import { Select } from '../ui/select';
-import { DatePicker } from '../ui/date-picker';
+import { Input } from '@/components/input';
+import { TextArea } from '@/components/text-area';
+import { ImageUpload } from '@/components/image-upload';
+import { Button } from '@/components/button';
+import { Select } from '@/components/select';
+import { DatePicker } from '@/components/date-picker';
 import { useEffect, useState } from 'react';
 import { CREATE_LEAD_STATUS, FORMITEMS } from '@/utils/constants/common-constants';
 import { FormItems } from '@/models/global-types';
